i need the app restared and then id like to see the landing page
